fixes issue with motor constructor

+++ lib/Motor/Motor.h
@@ -2,24 +2,27 @@
 #include <CurrentSense.h>
 #include <PID_v1.h>
 #include <pins.h>
+
 #ifndef MOTOR_H
 #define MOTOR_H
 #define RESISTANCE .047
+
 class Motor {
  private:
   CurrentSense current_sense;
-  float current;
-  float voltage;
-  float measured_power;
-  float set_point;
-  float out_power;
-  float kp;
-  float ki;
-  float kd;
+  double current;
+  double voltage;
+  double measured_power;
+  double set_point;
+  double out_power;
+  double kp;
+  double ki;
+  double kd;
 
  public:
   Motor(byte output_pin, byte current_pin, byte voltage_pin);
-  PID pid_obj;                        // I moved it to public so we can access pidObj.Compute() from outside
+  PID pid_obj;  // I moved it to public so we can access pidObj.Compute() from
+                // outside
   void update_power();                // Change mesuredPower
   void set_power(float pedal_power);  // Change setpoint
   float get_power();                  // Get current power setting
